Moved ExecutableMessage up one package to remove cyclic dependency.  Fairly large number of modifications here.
