+++ jenjin-core/src/main/java/com/jenjinstudios/core/message/ExecutablePingResponse.java
@@ -1,6 +1,7 @@
 package com.jenjinstudios.core.message;
 
 import com.jenjinstudios.core.Connection;
+import com.jenjinstudios.core.ExecutableMessage;
 import com.jenjinstudios.core.io.Message;
 
 /**
@@ -10,28 +11,28 @@ import com.jenjinstudios.core.io.Message;
  */
 public class ExecutablePingResponse extends ExecutableMessage
 {
-	private final Connection connection;
+    private final Connection connection;
 
-	/**
-	 * Construct a new {@code PingResponse}.
-	 *
-	 * @param connection The connection invoking this executable message.
-	 * @param message The message which caused this executable message to be invoked.
-	 */
-	public ExecutablePingResponse(Connection connection, Message message) {
-		super(message);
+    /**
+     * Construct a new {@code PingResponse}.
+     *
+     * @param connection The connection invoking this executable message.
+     * @param message The message which caused this executable message to be invoked.
+     */
+    public ExecutablePingResponse(Connection connection, Message message) {
+        super(message);
 
-		this.connection = connection;
-	}
+        this.connection = connection;
+    }
 
-	@Override
-	public void runDelayed() {
-		long requestTime = (long) getMessage().getArgument("requestTimeMillis");
+    @Override
+    public void runDelayed() {
+        long requestTime = (long) getMessage().getArgument("requestTimeMillis");
         connection.getPingTracker().addPingTime(System.currentTimeMillis() - requestTime);
     }
 
-	@Override
-	public void runImmediate() {
+    @Override
+    public void runImmediate() {
 
-	}
+    }
 }
