+++ jenjin-core/src/main/java/com/jenjinstudios/core/message/ExecutablePublicKeyMessage.java
@@ -1,6 +1,7 @@
 package com.jenjinstudios.core.message;
 
 import com.jenjinstudios.core.Connection;
+import com.jenjinstudios.core.ExecutableMessage;
 import com.jenjinstudios.core.io.Message;
 
 import java.net.InetAddress;
