+++ jenjin-core/src/test/java/com/jenjinstudios/core/message/DisabledExecutableMessage.java
@@ -1,6 +1,7 @@
 package com.jenjinstudios.core.message;
 
 import com.jenjinstudios.core.Connection;
+import com.jenjinstudios.core.ExecutableMessage;
 import com.jenjinstudios.core.io.Message;
 
 import java.util.logging.Level;
@@ -13,22 +14,22 @@ import java.util.logging.Logger;
  */
 public class DisabledExecutableMessage extends ExecutableMessage
 {
-	private static final Logger LOGGER = Logger.getLogger(DisabledExecutableMessage.class.getName());
+    private static final Logger LOGGER = Logger.getLogger(DisabledExecutableMessage.class.getName());
 
-	/**
-	 * Cosntruct a new {@code DisabledExecutableMessage}; not meant to be used in production code.
-	 *
-	 * @param connection The connection invoking this executable message.
-	 * @param message The message that caused this executable message to be invoked.
-	 */
-	public DisabledExecutableMessage(Connection connection, Message message) {
-		super(message);
-		LOGGER.log(Level.SEVERE, "DisabledExecutableMessage constructed in {0}", connection);
-	}
+    /**
+     * Cosntruct a new {@code DisabledExecutableMessage}; not meant to be used in production code.
+     *
+     * @param connection The connection invoking this executable message.
+     * @param message The message that caused this executable message to be invoked.
+     */
+    public DisabledExecutableMessage(Connection connection, Message message) {
+        super(message);
+        LOGGER.log(Level.SEVERE, "DisabledExecutableMessage constructed in {0}", connection);
+    }
 
-	@Override
-	public void runDelayed() { throw new IllegalStateException("This message should be disabled."); }
+    @Override
+    public void runDelayed() { throw new IllegalStateException("This message should be disabled."); }
 
-	@Override
-	public void runImmediate() { throw new IllegalStateException("This message should be disabled."); }
+    @Override
+    public void runImmediate() { throw new IllegalStateException("This message should be disabled."); }
 }
