+++ jenjin-core/src/test/java/com/jenjinstudios/core/message/ExecutableBlankMessage.java
@@ -1,6 +1,7 @@
 package com.jenjinstudios.core.message;
 
 import com.jenjinstudios.core.Connection;
+import com.jenjinstudios.core.ExecutableMessage;
 import com.jenjinstudios.core.io.Message;
 
 import java.util.logging.Level;
@@ -13,28 +14,28 @@ import java.util.logging.Logger;
  */
 public class ExecutableBlankMessage extends ExecutableMessage
 {
-	private static final Logger LOGGER = Logger.getLogger(ExecutableBlankMessage.class.getName());
-	private final Connection connection;
+    private static final Logger LOGGER = Logger.getLogger(ExecutableBlankMessage.class.getName());
+    private final Connection connection;
 
-	/**
-	 * Construct a new ExecutableBlankMessage; not meant to be used in production code.
-	 *
-	 * @param connection The connection invoking this message.
-	 * @param message The message causing this to be invoked.
-	 */
-	@SuppressWarnings("WeakerAccess")
-	public ExecutableBlankMessage(Connection connection, Message message) {
-		super(message);
-		this.connection = connection;
-	}
+    /**
+     * Construct a new ExecutableBlankMessage; not meant to be used in production code.
+     *
+     * @param connection The connection invoking this message.
+     * @param message The message causing this to be invoked.
+     */
+    @SuppressWarnings("WeakerAccess")
+    public ExecutableBlankMessage(Connection connection, Message message) {
+        super(message);
+        this.connection = connection;
+    }
 
-	@Override
-	public void runDelayed() {
+    @Override
+    public void runDelayed() {
 
-	}
+    }
 
-	@Override
-	public void runImmediate() {
-		LOGGER.log(Level.FINEST, "{0} received blank message.", connection.getName());
-	}
+    @Override
+    public void runImmediate() {
+        LOGGER.log(Level.FINEST, "{0} received blank message.", connection.getName());
+    }
 }
